+++ src/main/java/net/andrecarbajal/mine_control_cli/model/github/GithubTagResponse.java
@@ -1,0 +1,10 @@
+package net.andrecarbajal.mine_control_cli.model.github;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class GithubTagResponse {
+    private String name;
+}