+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Fix JavaPathPropertyValidator to allow java as a valid path
+- Check updates on startup and notify user if there is a new version available